+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 - `PostgreSQLContainerConfig`:
   - `postgresPort`, `maxConnections`, `logStatement`.
 
+- `MySQLContainerConfig`:
+  - `_buildImageArgs`: Ignore `forceNativePasswordAuthentication` for version 9+ (and latest).
+
 - swiss_knife: ^3.2.2
 - mercury_client: ^2.2.4
 - logging: ^1.3.0